% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contourfinder.R
\name{contourfinder}
\alias{contourfinder}
\title{Find and Highlight Contours}
\usage{
contourfinder(
  ref_image,
  contourvalue,
  rgbcolor = c(1, 0, 0),
  save = FALSE,
  output_ref_image = NA
)
}
\arguments{
\item{ref_image}{The name of the input image}

\item{contourvalue}{The minimum pixel value you want to keep (a low contour value captures weaker contrasts)}

\item{rgbcolor}{A vector containing the rgb values for the color mask on a 0-1 scale. Default is c(1,0,0), which is pure red.}

\item{save}{Logical. Save the image as a new file. Default is FALSE.}

\item{output_ref_image}{If save == TRUE, the name you want to give the output image.}
}
\value{
A recolored image. If save == TRUE, will save the recolored image to your working directory
}
\description{
Define a rectangular region and locate pixels with contour value greater than the one defined within the function. Then, recolor pixels and display recolored image.
}
