% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_apply.R
\name{ct_apply}
\alias{ct_apply}
\title{Overlay Contours Onto Images}
\usage{
ct_apply(
  images,
  ref_images = 1,
  roi_in = NULL,
  contour_vals = 0.1,
  regions = 1,
  shift = c(0, 0),
  color = "red",
  show_image = TRUE,
  process = "sequential",
  cores = 1
)
}
\arguments{
\item{images}{A vector containing file paths of images to be analyzed. The first image will be the reference image, and contours from the first image will be superimposed onto the other images.}

\item{ref_images}{A numeric indicating in how many reference images you want to search for contours. Default is 1. When greater than 1, the first n images will be used.}

\item{roi_in}{An argument for delineating the region(s) of interest outside of the main function. Default is NULL and will launch a user interface so the user can draw a region(s) of interest on the image. If "roi_in" is not NULL, input should be a list of length ref_images, and each item of the list should be a separate 4-column data frame for the region(s) of interest in each image. The data frame needs to contain coordinates to the region(s) of interest in the following order: top-left x, top-left y, bottom-right x, bottom-right y. Note that plotting of images by "imager" starts in the top-left corner.}

\item{contour_vals}{A vector of numeric values between 0-1 for the lowest contour value you want to recolor in each image(a low contour value captures weaker contrasts). Default is 0.1.}

\item{regions}{A numeric indicating how many regions to draw. Default is 1.}

\item{shift}{A vector of length 2 containing numerics indicating the amount of shift along the x axis first and the y axis second. Positive values indicate shifts right or up, while negative values indicate shifts left or down.}

\item{color}{A character string for the color of the superimposed object. Default is red.}

\item{show_image}{Logical. Plot images as they are recolored. Default is TRUE.}

\item{process}{One of "sequential" or "parallel" to indicate how images should be processed. Default is sequential. If "parallel", cores will also have to be set.}

\item{cores}{A numeric indicating how many cores to use in parallel processing. Safest is availableCores() - 1. Default is 1.}
}
\value{
Recolored images. Saves recolored images to working directory with "_contourr" appended to the original name.
}
\description{
Find pixel coordinates of contours in defined regions of a reference image. Then, find and recolor the same pixels in new images.
}
