% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_display.R
\name{ct_display}
\alias{ct_display}
\title{Display an Image}
\usage{
ct_display(cimage)
}
\arguments{
\item{cimage}{An image to display.}
}
\value{
An image.
}
\description{
Display an image in the plot pane.
}
