% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_im_to_df.R
\name{ct_im_to_df}
\alias{ct_im_to_df}
\title{Convert an Image Into a Data Frame}
\usage{
ct_im_to_df(cimage)
}
\arguments{
\item{cimage}{The image to be converted.}
}
\value{
The input image as a data frame, with a new column for the unique identifier of each pixel.
}
\description{
Convert an image into a data frame with a new unique identifier column to be used in the contourr workflow.
}
