% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_match.R
\name{ct_match}
\alias{ct_match}
\title{Find the Pixels to be Recolored in a Raw Image}
\usage{
ct_match(imdf, roi, shift = c(0, 0))
}
\arguments{
\item{imdf}{A data frame of the raw image. The data frame must contain a column with unique identifiers for each pixels. See ct_im_to_df.}

\item{roi}{A data frame containing the locations of the region of interest. See ct_define_roi.}

\item{shift}{A vector of length 2 containing numerics indicating the amount of shift along the x axis first and the y axis second. Positive values indicate shifts right or up, while negative values indicate shifts left or down. Default is no shift.}
}
\value{
The input data frame as an image.
}
\description{
Identify pixels that are contour pixels in the raw image to be recolored later. Optionally shift the pixels that are recolored.
}
