% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_name.R
\name{ct_name}
\alias{ct_name}
\title{Find New Name to be Given to Recolored Image}
\usage{
ct_name(imagepath)
}
\arguments{
\item{imagepath}{The raw image name as a character.}
}
\value{
The new contourr image name (the raw file name with "_contourr" appended before the filename extension).
}
\description{
Append "_contourr" to the given filename
}
