\name{de4DGE}
\alias{de4DGE}

\title{Compute Moderated Differential Expression Scores for Digital Gene Expression (DGE) Data}

\description{Runs weighted likelihood calculation for moderated estimates of dispersion, and tests for differences in 'tag' abundance between groups}

\usage{ 
de4DGE(object,prior.n=10, disp.init=NULL, doPoisson=FALSE, useCommonDisp=TRUE, verbose=TRUE) 
}

\arguments{ 
\item{object}{\code{DGEList} object containing (at least) elements \code{counts} (matrix: rows-tags/genes, columns-libraries), \code{lib.size}, \code{group} indicating class} 

\item{prior.n}{numeric scalar for the smoothing parameter that indicates the weight to put on the common likelihood compared to the individual tag's likelihood; default \code{10} means that the common likelihood is given 10 times the weight of the individual tag/gene's likelihood in the estimation of the tag/genewise dispersion}

\item{disp.init}{initialized value(s) of the dispersion parameter, can be a common value or tag/genewise values; if \code{NULL}, then the common value on Poisson-adjusted counts is used}

\item{doPoisson}{logical, if \code{TRUE} then use a Poisson model rather than Negative Binomial to analyse the data; default \code{FALSE}}

\item{useCommonDisp}{logical, if \code{TRUE} then the common dispersion estimate is used for all tags/genes, otherwise tag/genewise dispersion parameters are estimated; default \code{TRUE}}

\item{verbose}{logical, whether to write comments, default \code{TRUE}}
}

\value{ \code{deDGEList} object with elements 
	\item{conc}{list containing concentration estimates}
	\item{dispersion}{estimates of dispersion parameter)}
	\item{pseudo}{numeric matrix of pseudocounts generated by \code{quantileAdjust}}
	\item{group}{vector or factor indicating the experimental class of each sample}
	\item{M}{numeric scalar giving the library size to which counts are adjusted; the geometric mean of the original library sizes}
}

\details{
An older function, no longer included in the recommended analysis pathway for DGE data. Instead, see \code{\link{estimateCommonDisp}}, \code{\link{estimateTagwiseDisp}} and \code{\link{exactTest}}.
}


\author{Mark Robinson, Davis McCarthy}

\examples{
# generate raw counts from NB, create list object
y<-matrix(rnbinom(20,size=1,mu=10),nrow=5)
d<-DGEList(counts=y,group=rep(1:2,each=2),lib.size=rep(c(1000:1001),2))

# estimate common dispersion, find smoothing parameter and call main procedure to find differences in expression between groups
d<-estimateCommonDisp(d)
prior.n<-estimateSmoothing(d)
ms<-de4DGE(d,prior.n=prior.n,disp.init=d$common.dispersion)
}

\references{
Robinson MD, Smyth GK. 'Small-sample estimation of negative binomial dispersion, with applications to SAGE data.' Biostatistics. 2008 Apr;9(2):321-32.

Robinson MD, Smyth GK. 'Moderated statistical tests for assessing differences in tag abundance.' Bioinformatics. 2007 Nov 1;23(21):2881-7.
}

\keyword{algebra}