\name{deDGE}
\alias{deDGE}

\title{Compute Moderated Differential Expression Scores for Digital Gene Expression (DGE) Data}

\description{Runs weighted likelihood calculation for moderated estimates of dispersion, and tests for differences in 'tag' abundance between groups}

\usage{ 
deDGE(object,alpha=500,doPoisson=FALSE,verbose=TRUE)
}

\arguments{ 
\item{object}{\code{DGEList} containing elements \code{counts} (matrix: rows-tags, columns-libraries), \code{lib.size}, \code{group} indicating class} 

\item{alpha}{weight to put on the individual tag's likelihood}

\item{doPoisson}{logical, whether to fit Poisson model instead of Negative Binomial, default \code{FALSE}}

\item{verbose}{logical, whether to write comments, default \code{TRUE}}
}

\value{ \code{deDGEList} with elements \code{ps} (list containing proportion estimates), \code{r} (estimates of 1/overdispersion), \code{pseudo} (pseudocounts generated by \code{quantileAdjust}), \code{group} (indicating class of each sample), \code{M} (geometric mean of library sizes)}

\details{
An older function, no longer included in the recommended analysis pathway for DGE data. Instead, see \code{\link{estimateCommonDisp}}, \code{\link{estimateTagwiseDisp}} and \code{\link{exactTest}}.
}


\author{Mark Robinson, Davis McCarthy}

\examples{
# generate raw counts from NB, create list object
y<-matrix(rnbinom(20,size=1,mu=10),nrow=5)
d<-DGEList(counts=y,group=rep(1:2,each=2),lib.size=rep(c(1000:1001),2))

# find alpha and call main procedure to find differences
alpha<-alpha.approxeb(d)
ms<-deDGE(d,alpha=alpha$alpha)
}

\references{
Robinson MD, Smyth GK. 'Small-sample estimation of negative binomial dispersion, with applications to SAGE data.' Biostatistics. 2008 Apr;9(2):321-32.

Robinson MD, Smyth GK. 'Moderated statistical tests for assessing differences in tag abundance.' Bioinformatics. 2007 Nov 1;23(21):2881-7.
}

\keyword{algebra}
