% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_plot.R
\name{edge_plot}
\alias{edge_plot}
\title{Plot an edge matching puzzle.}
\usage{
edge_plot(tiles)
}
\arguments{
\item{tiles}{List of character vectors of the form
list(c("red", "yellow", "blue"), ..., c("white", "black", "green")) of
length 16. This list represents the colours of the dots on 16 equilateral
triangle tiles that will tessalate into a larger triangle.
\code{tiles} should start with the bottom left tile and end with the
top tile, moving left to right, row by row. The colours within each
character vector go clock-wise around each triangle, starting with the
top left dot for triangles pointing upward, and the bottom right dot for
triangles pointing downward.}
}
\value{
A ternary diagram displaying the puzzle
}
\description{
Plot an edge matching puzzle.
}
\examples{
edge_plot(tiles = rep(list(c("black", "white", "blue")), 16))
}
