% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_try.R
\name{edge_try}
\alias{edge_try}
\title{Try fitting as many candidate tiles onto the given attempt as possible}
\usage{
edge_try(attempt, candidates)
}
\arguments{
\item{attempt}{List of tiles which have been placed}

\item{candidates}{List of tiles which haven't been placed yet}
}
\value{
List of two lists names attempt and discard. attempt is a list of
tiles that is at least as long as the input parameter \code{attempt}. discard
is the list of tiles which could not be fitted onto \code{attempt}.
}
\description{
Try fitting as many candidate tiles onto the given attempt as possible
}
\examples{

}
