% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_df_to_im_bw.R
\name{edger_df_to_im_bw}
\alias{edger_df_to_im_bw}
\title{Convert a Data Frame into an Image}
\usage{
edger_df_to_im_bw(imdf, show = FALSE)
}
\arguments{
\item{imdf}{A data frame to be converted. Must contain at least 3 columns: x, y, and value. If a color image, must also contain a cc column.}

\item{show}{A Boolean indicating if the new image should be displayed. Default is FALSE.}
}
\value{
The input data frame as an image.
}
\description{
Convert a data frame into either a black and white or a color image. Optionally display the new image.
}
