% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_display.R
\name{edger_display}
\alias{edger_display}
\title{Display an Image}
\usage{
edger_display(cimage)
}
\arguments{
\item{cimage}{An image to display.}
}
\value{
An image.
}
\description{
Display an image in the plot pane.
}
