% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_match.R
\name{edger_match}
\alias{edger_match}
\title{Find the Pixels to be Recolored in a Raw Image}
\usage{
edger_match(imdf, roi, shift = c(0, 0), rotate = 0)
}
\arguments{
\item{imdf}{A data frame of the raw image. The data frame must contain a column with unique identifiers for each pixel. See edger_im_to_df.}

\item{roi}{A data frame containing the locations of the contour pixels in the region of interest. See edger_define_roi.}

\item{shift}{A vector of length 2 containing numerics indicating the amount to shift the recolored pixels along the x axis first and the y axis second. Positive values indicate shifts right or up, while negative values indicate shifts left or down. Default is no shift.}

\item{rotate}{A numeric indicating the number of degrees to rotate the recolored pixels. Pixels are rotated around the center of the contour pixels.}
}
\value{
The input data frame as an image.
}
\description{
Identify pixels that are contour pixels in the raw image to be recolored later. Optionally shift and/or rotate the pixels that are recolored.
}
