% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_meta.R
\name{edger_meta}
\alias{edger_meta}
\title{Extract and Reattribute Metadata}
\usage{
edger_meta(imagepaths, tags)
}
\arguments{
\item{imagepaths}{The file path to the original image}

\item{tags}{The Exiftool tags for metadata to extract. The default is all tags.}
}
\value{
Metadata attributed to the recolored images.
}
\description{
Attribute metadata from original images to recolored images. Doesn't work yet...
}
