% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_name.R, R/edger_name_old.R
\name{edger_name}
\alias{edger_name}
\title{Find New Name to be Given to Recolored Image}
\usage{
edger_name(imagepath)

edger_name(imagepath)
}
\arguments{
\item{imagepath}{The raw image name as a character.}
}
\value{
The new edger image name (the raw file name with "_edger" appended before the filename extension).

The new contourr image name (the raw file name with "_contourr" appended before the filename extension).
}
\description{
Append "_edger" to the given filename

Append "_contourr" to the given filename
}
