% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_overlay.R
\name{edger_overlay}
\alias{edger_overlay}
\title{Apply Contours to an Image}
\usage{
edger_overlay(imagepath, m, rgbcolor, show_image = FALSE)
}
\arguments{
\item{imagepath}{The file path to the image to be recolored.}

\item{m}{A vector of unique identifiers to pixels that were contours in a different image.}

\item{rgbcolor}{A vector of new values for each of the color channels.}

\item{show_image}{A Boolean indicating if the image should be displayed. Default is FALSE.}
}
\value{
Recolors and saves image.
}
\description{
After finding contours in image(s), apply those contours to a new image.
}
