% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_recolor.R
\name{edger_recolor}
\alias{edger_recolor}
\title{Find and Recolor Pixels in an Image}
\usage{
edger_recolor(imdf, m, rgbcolor, show = TRUE)
}
\arguments{
\item{imdf}{A color image as a data frame.}

\item{m}{A vector containing unique identifiers of pixels to be recolored.}

\item{rgbcolor}{A vector of the new values for each color channel.}

\item{show}{A Boolean indicating if the new image should be shown. Default is TRUE.}
}
\value{
A new image with pixels in m recolored.
}
\description{
Recolor pixels contained in m. Optionally display recolored image.
}
