% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_roi.R
\name{edger_roi}
\alias{edger_roi}
\title{Find Contour Values for Several Images}
\usage{
edger_roi(imagepaths, regions = 1, color = "red")
}
\arguments{
\item{imagepaths}{A vector containing file paths of images to be analyzed.}

\item{regions}{A numeric indicating how many regions to draw. Default is 1.}

\item{color}{A character string for the color of the superimposed object. Default is red.}
}
\value{
A list of lists, with each list containing the image file path and a data frame of coordinates to region(s) of interest and contour values selected for that image.
}
\description{
An interactive function for locating pixels corresponding to edges in multiple regions and multiple photos. If the default contour value (0.1) does not display desired contours, input "N" when prompted to input a new contour value and redisplay. Repeat as necessary until it is satisfactory, then input "Y". The drawn region(s) of interest and contour values will be output as a list of lists, with one list per image.
}
