% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_save.R
\name{edger_save}
\alias{edger_save}
\title{Save an Image}
\usage{
edger_save(cimage, imagename)
}
\arguments{
\item{cimage}{The image to be saved.}

\item{imagename}{A name to give the image.}
}
\value{
Saves image
}
\description{
Save an image with the given filename.
}
