% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_testr.R
\name{edger_testr}
\alias{edger_testr}
\title{Interactively Choose Thresholds, Shift, and Rotation}
\usage{
edger_testr(imagepaths, regions = 1, color = "red")
}
\arguments{
\item{imagepaths}{A vector containing file paths of images to be analyzed.}

\item{regions}{A numeric indicating how many regions to draw. Default is 1.}

\item{color}{A character string for the color of the superimposed object. Default is red.}
}
\value{
A list of coordinates to region(s) of interest, edge values, shifts, and rotations for each image. The roi and shift list items are output as lists, while th and rotate are output as vectors.
}
\description{
An interactive function for locating pixels corresponding to edges in multiple regions and multiple photos. If the default threshold value (0.1) does not display desired contours, input "N" when prompted to input a new threshold value and redisplay. Repeat as necessary until it is satisfactory, then input "Y". Do the same procedure for shift and rotation. The drawn region(s) of interest, threshold values, shifts, and rotations will be output as a list.
}
