% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_edge.R
\name{match_edge}
\alias{match_edge}
\title{Check if a tile matches onto a given permutation of tiles.}
\usage{
match_edge(tiles, tile)
}
\arguments{
\item{tiles}{List of character vectors from the \code{edge_plot()}
parameter \code{tiles}.}

\item{tile}{character vector in the form of a element of \code{tiles}.}
}
\value{
Boolean indicating whether \code{tile} will fit onto \code{tiles}.
}
\description{
Check if a tile matches onto a given permutation of tiles.
}
\examples{
match_edge(tiles = rep(list(c("black", "white", "blue")), 15),
           tile = c("black", "white", "blue"))
}
