% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_tile.R
\name{match_tile}
\alias{match_tile}
\title{Check if a tile (in any of the three rotations) matches onto a given
permutation of tiles.}
\usage{
match_tile(tiles, tile)
}
\arguments{
\item{tiles}{List of character vectors from the \code{edge_plot()}
parameter \code{tiles}.}

\item{tile}{character vector in the form of a element of \code{tiles}.}
}
\value{
If \code{tile} matches on to \code{tiles} in some rotation,
\code{TRUE}. Otherwise, \code{FALSE}.
}
\description{
Check if a tile (in any of the three rotations) matches onto a given
permutation of tiles.
}
