% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{overlay}
\alias{overlay}
\title{Overlay Contours Onto Images}
\usage{
overlay(
  ref_image,
  contourvalue,
  rgbcolor = c(1, 0, 0),
  regions = 1,
  output_ref_image,
  image_names,
  output_image_names
)
}
\arguments{
\item{ref_image}{The name of the reference image}

\item{contourvalue}{The minimum pixel value you want to keep (a low contour value captures weaker contrasts)}

\item{rgbcolor}{A vector containing the RGB values for the color mask on a 0-1 scale. Default is c(1,0,0), which is pure red}

\item{regions}{A numeric indicating how many regions to draw. Default is 1}

\item{output_ref_image}{The name you want to give the output reference image.}

\item{image_names}{a vector containing the names of the images you want to recolor.}

\item{output_image_names}{a vector containing the names of the recolored images. Must be the same length as image_names.}
}
\value{
Recolored images. Saves recolored images to your working directory
}
\description{
Find pixel coordinates of contours in defined regions of a reference image. Then, find and recolor the same pixels in new images.
}
