% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_tile.R
\name{put_tile}
\alias{put_tile}
\title{Transform a tile.}
\usage{
put_tile(tile_in, scale = 0.25, reflect = FALSE, x = 0, y = 0)
}
\arguments{
\item{tile_in}{Data.frame with two columns (x and y) and three rows (one for
each dot)}

\item{scale}{Numeric between 0 and 1}

\item{reflect}{Boolean specifying whether to relect tile in the y axis (i.e.
turn upside down)}

\item{x}{Numeric between 0 and 0.75 for horizontal translation}

\item{y}{Numeric between 0 and 3^.5/2 (the height of an equilateral triangle
of unit side length) for vertical translation}
}
\value{
A transformed tile
}
\description{
Transform a tile.
}
\examples{
r <- (2*3^.5+4)^-1
h <- r * (3^.5+1)
big_tile <- data.frame(x = c(.5-r, .5, .5+r),
                       y = c(h, r, h))

put_tile(tile = big_tile, reflect = TRUE, x = .125)
}
