% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_tiles.R
\name{put_tiles}
\alias{put_tiles}
\title{Turn a list of coloured tiles into a dataframe ready for plotting.}
\usage{
put_tiles(tiles_in)
}
\arguments{
\item{tiles_in}{List of character vectors from the \code{edge_plot()}
parameter \code{tiles}.}
}
\value{
Data.frame with the xy coordinates and colour for all of the dots for
    all 16 tiles.
}
\description{
Turn a list of coloured tiles into a dataframe ready for plotting.
}
\examples{
put_tiles(list(c("green", "red", "yellow")
,c("green", "yellow", "white"),
c("white", "white", "blue")
,c("blue", "black", "white")
,c("white", "green", "red")
,c("black", "green", "red")
,c("green", "black", "black")
,c("yellow", "green", "white")
,c("blue", "yellow", "red")
,c("green", "red", "black")
,c("blue", "white", "blue")
,c("white", "black", "blue")
,c("red", "yellow", "white")
,c("yellow", "black", "blue")
,c("green", "yellow", "black")
,c("green", "black", "red")))
}
