% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_find.R
\name{ct_find}
\alias{ct_find}
\title{Find and Recolor Contours in a Single Image}
\usage{
ct_find(image, contour_value = 0.1, color = "red", regions = 1, save = FALSE)
}
\arguments{
\item{image}{The file path to the image to be analyzed.}

\item{contour_value}{A numeric between 0-1 for the lowest contour value you want to recolor (a low contour value captures weaker contrasts). Default is 0.1.}

\item{color}{A character string for the color of the superimposed object. Default is red.}

\item{regions}{A numeric indicating how many regions to draw. Default is 1.}

\item{save}{Logical. Save the output image. Default is FALSE.}
}
\value{
A recolored image. If save == TRUE, saves recolored image to working directory with "_contourr" appended to the original name.
}
\description{
Define a rectangular region and locate pixels with contour value greater than the one defined within the function. Then, recolor pixels and display recolored image.
}
