% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_find.R
\name{ct_find}
\alias{ct_find}
\title{Find and Highlight Contours}
\usage{
ct_find(image, contourvalue = 0.1, color = "red", regions = 1, save = FALSE)
}
\arguments{
\item{image}{The name of the input image.}

\item{contourvalue}{The minimum pixel value you want to keep (a low contour value captures weaker contrasts). Default is 0.1.}

\item{color}{A character string for the color of the superimposed object. Default is red.}

\item{regions}{A numeric indicating how many regions to draw. Default is 1.}

\item{save}{Logical. Save the output image. Default is FALSE.}
}
\value{
A recolored image. If save == TRUE, saves recolored image to a "contourr_images" directory in your working directory.
}
\description{
Define a rectangular region and locate pixels with contour value greater than the one defined within the function. Then, recolor pixels and display recolored image.
}
