% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_overlay.R
\name{ct_overlay}
\alias{ct_overlay}
\title{Overlay Contours Onto Images}
\usage{
ct_overlay(images, contourvalue = 0.1, color = "red", regions = 1)
}
\arguments{
\item{images}{A vector containing image paths. The first image will be the reference image, and contours from the first image will be superimposed onto the other images.}

\item{contourvalue}{The minimum pixel value you want to keep (a low contour value captures weaker contrasts). Default is 0.1.}

\item{color}{A character string for the color of the superimposed object. Default is red.}

\item{regions}{A numeric indicating how many regions to draw. Default is 1.}
}
\value{
Recolored images. Saves recolored images to a "contourr_images" directory in your working directory.
}
\description{
Find pixel coordinates of contours in defined regions of a reference image. Then, find and recolor the same pixels in new images.
}
