% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edger_meta.R
\name{edger_meta}
\alias{edger_meta}
\title{Extract and Reattribute Metadata}
\usage{
edger_meta(imagepath)
}
\arguments{
\item{imagepath}{The file path to the original image. Original and recolored image must be in the same directory.}
}
\value{
Recolored image with metadata tags from original image
}
\description{
Attribute metadata from original image to recolored image. Requires that ExifTool is installed on your machine and is in your environmental variables.
}
