% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{overlay}
\alias{overlay}
\title{Overlay Contours Onto Images}
\usage{
overlay(
  ref_image,
  contourvalue,
  rgbcolor = c(1, 0, 0),
  output_ref_image,
  image_names,
  output_image_names,
  imwidth,
  imht
)
}
\arguments{
\item{ref_image}{The name of the input reference image}

\item{contourvalue}{The minimum pixel value you want to keep (a low contour value captures weaker contrasts)}

\item{rgbcolor}{A vector containing the RGB values for the color mask on a 0-1 scale. Default is c(1,0,0), which is pure red.}

\item{output_ref_image}{The name you want to give the output reference image.}

\item{image_names}{a vector containing the names of the photos you want to recolor.}

\item{output_image_names}{a vector containing the names for the new images with recolored pixels. Must be the same length as image_names.}

\item{imwidth}{Image width, in pixels.}

\item{imht}{Image height, in pixels .}
}
\value{
Recolored images. Will save the recolored image to your working directory
}
\description{
Define a rectangular region and locate pixels with contour value greater than the one defined within the function. Then, recolor pixels and display recolored image.
}
