\name{Tu102}
\alias{Tu102}
\alias{Tu98}
\alias{NC1}
\alias{NC2}
\docType{data}
\title{Raw Data for Several SAGE Libraries from the Zhang 1997 Science Paper.}
\description{
  SAGE dataset for 2 tumour samples, 2 normal samples.
}
\usage{data(Tu102)}
\format{
  Data frames with 22713, 18794, 16270 and 17703 observations (for Tu102, Tu98, NC2, NC1, respectively) on the following 2 variables.
  \describe{
    \item{\code{Tag_Sequence}}{a character vector}
    \item{\code{Count}}{a numeric vector}
  }
}
\source{
Zhang et al. (1997) Gene Expression Profiles in Normal and Cancer Cells. \emph{Science}, 276, 1268-72.
}
\keyword{datasets}
